use X macros to serialize the list of variables in ECS components `submodules/liborangepower`: `git pull`

+++ src/game_engine/playfield_layer_prio_enum.hpp
@@ -25,7 +25,7 @@ namespace dungwich_sandeon
 namespace game_engine
 {
 
-enum class PlayfieldLayerPrio: size_t
+enum class PlayfieldLayerPrio: uint
 {
 	Block,
 	Item,
